FIX: require cycle errors
