+++ class/hd-segwit-bech32-transaction.js
@@ -1,4 +1,5 @@
-import { HDSegwitBech32Wallet, SegwitBech32Wallet } from './';
+import { HDSegwitBech32Wallet } from './wallets/hd-segwit-bech32-wallet';
+import { SegwitBech32Wallet } from './wallets/segwit-bech32-wallet';
 const bitcoin = require('bitcoinjs-lib');
 const BlueElectrum = require('../blue_modules/BlueElectrum');
 const reverse = require('buffer-reverse');
