+++ class/index.js
@@ -1,5 +1,5 @@
-export * from './wallets/abstract-wallet';
 export * from './app-storage';
+export * from './wallets/abstract-wallet';
 export * from './wallets/legacy-wallet';
 export * from './wallets/segwit-bech32-wallet';
 export * from './wallets/segwit-p2sh-wallet';
@@ -10,8 +10,8 @@ export * from './wallets/watch-only-wallet';
 export * from './wallets/lightning-custodian-wallet';
 export * from './wallets/abstract-hd-wallet';
 export * from './wallets/hd-segwit-bech32-wallet';
-export * from './hd-segwit-bech32-transaction';
 export * from './wallets/placeholder-wallet';
 export * from './wallets/hd-legacy-electrum-seed-p2pkh-wallet';
 export * from './wallets/hd-segwit-electrum-seed-p2wpkh-wallet';
 export * from './wallets/multisig-hd-wallet';
+export * from './hd-segwit-bech32-transaction';
